% Auto-generated: do not edit by hand
\name{dtcAsyncComponent}

\alias{dtcAsyncComponent}

\title{AsyncComponent component}

\description{

}

\usage{
dtcAsyncComponent(id=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{value}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

