% Auto-generated: do not edit by hand
\name{dtcCollapseComponent}

\alias{dtcCollapseComponent}

\title{CollapseComponent component}

\description{

}

\usage{
dtcCollapseComponent(children=NULL, display=NULL, id=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{display}{Logical. }

\item{id}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

