% Auto-generated: do not edit by hand
\name{dtcDelayedEventComponent}

\alias{dtcDelayedEventComponent}

\title{DelayedEventComponent component}

\description{

}

\usage{
dtcDelayedEventComponent(id=NULL, n_clicks=NULL)
}

\arguments{
\item{id}{Character. }

\item{n_clicks}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

