% Auto-generated: do not edit by hand
\name{dtcFragmentComponent}

\alias{dtcFragmentComponent}

\title{FragmentComponent component}

\description{

}

\usage{
dtcFragmentComponent(children=NULL, id=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

