% Auto-generated: do not edit by hand
\name{dtcStyledComponent}

\alias{dtcStyledComponent}

\title{StyledComponent component}

\description{
MyComponent description
}

\usage{
dtcStyledComponent(id=NULL, style=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The id of the component}

\item{style}{The style}

\item{value}{Character. The value to display}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

