% Auto-generated: do not edit by hand
\name{dtcWidthComponent}

\alias{dtcWidthComponent}

\title{WidthComponent component}

\description{

}

\usage{
dtcWidthComponent(id=NULL, width=NULL)
}

\arguments{
\item{id}{Character. }

\item{width}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

